ai-model/requirements.txt
# Core data processing and machine learning
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Deep learning framework
tensorflow>=2.10.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# HTTP requests for data downloading
requests>=2.28.0

# Configuration file handling
PyYAML>=6.0

# Additional utilities
tqdm>=4.64.0  # Progress bars
joblib>=1.2.0  # Parallel processing and model persistence

# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0  # Code formatting
flake8>=5.0.0  # Code linting
mypy>=0.991  # Type checking

# Optional: Jupyter notebook support for analysis
jupyter>=1.0.0
ipykernel>=6.15.0

# Optional: Advanced visualization
plotly>=5.10.0

# Optional: Model interpretability
shap>=0.41.0

# Optional: Hyperparameter optimization
optuna>=3.0.0

# Web framework for serving the application
fastapi==0.99.1
uvicorn[standard]==0.23.2
openai-agents==0.3.0
python-dotenv==1.0.0